requirements.txt
streamlit
torch
sentence-transformers
transformers
datasets
Pillow
openai